# pkg/logentry: refactor stages to use channels