--- a/pkg/promtail/targets/filetargetmanager.go
+++ b/pkg/promtail/targets/filetargetmanager.go
@@ -108,6 +108,9 @@ func NewFileTargetManager(
 			}
 		}
 
+		// Run the pipeline
+		pipeline.Start(ctx)
+
 		s := &targetSyncer{
 			log:            logger,
 			positions:      positions,