--- a/pkg/promtail/targets/journaltargetmanager_linux.go
+++ b/pkg/promtail/targets/journaltargetmanager_linux.go
@@ -3,6 +3,8 @@
 package targets
 
 import (
+	"context"
+
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/grafana/loki/pkg/logentry/stages"
@@ -16,6 +18,7 @@ import (
 type JournalTargetManager struct {
 	logger  log.Logger
 	targets map[string]*JournalTarget
+	quit    func()
 }
 
 // NewJournalTargetManager creates a new JournalTargetManager.
@@ -25,9 +28,12 @@ func NewJournalTargetManager(
 	client api.EntryHandler,
 	scrapeConfigs []scrape.Config,
 ) (*JournalTargetManager, error) {
+	ctx, quit := context.WithCancel(context.Background())
+
 	tm := &JournalTargetManager{
 		logger:  logger,
 		targets: make(map[string]*JournalTarget),
+		quit:    quit,
 	}
 
 	for _, cfg := range scrapeConfigs {
@@ -40,6 +46,9 @@ func NewJournalTargetManager(
 		if err != nil {
 			return nil, err
 		}
+
+		// Run the pipeline
+		pipeline.Start(ctx)
 
 		t, err := NewJournalTarget(
 			logger,
@@ -76,6 +85,8 @@ func (tm *JournalTargetManager) Stop() {
 			level.Error(t.logger).Log("msg", "error stopping JournalTarget", "err", err.Error())
 		}
 	}
+
+	tm.quit()
 }
 
 // ActiveTargets returns the list of JournalTargets where journal data