--- a/pkg/promtail/targets/journaltargetmanager_linux.go
+++ b/pkg/promtail/targets/journaltargetmanager_linux.go
@@ -47,12 +47,9 @@ func NewJournalTargetManager(
 			return nil, err
 		}
 
-		// Run the pipeline
-		pipeline.Start(ctx)
-
 		t, err := NewJournalTarget(
 			logger,
-			pipeline.Wrap(client),
+			pipeline.Wrap(ctx, client),
 			positions,
 			cfg.JobName,
 			cfg.RelabelConfigs,