Proof of concept, do not merge!

This commit introduces an AsyncStage interface to run a Stage as an long-lasting process, manipulating log lines through channels. Existing stages have been wrapped as asynchronous stages.

An alternative solution using callbacks is in #1373.

/cc @slim-bean @cyriltovena