@rfratto 2019-12-05
Some rough benchmarking shows that this is (obviously) slower than the existing code across the board, since there's no more concurrency from multiple FileTargets. I simulated sending 1000 logs concurrently from 1, 10, 100, and 1000 clients. These are the results:

Old code:

```
BenchmarkPipeline_Concurrency/1_clients-8                  10000               100 ns/op
BenchmarkPipeline_Concurrency/10_clients-8                 10000               760 ns/op
BenchmarkPipeline_Concurrency/100_clients-8                10000              5217 ns/op
BenchmarkPipeline_Concurrency/1000_clients-8               10000             56157 ns/op
```

New code:

```
BenchmarkPipeline_Concurrency/1_clients-8                  10000               475 ns/op
BenchmarkPipeline_Concurrency/10_clients-8                 10000              4664 ns/op
BenchmarkPipeline_Concurrency/100_clients-8                10000             46876 ns/op
BenchmarkPipeline_Concurrency/1000_clients-8               10000            261081 ns/op
```